+++ yzt_service/src/main/java/com/maizhiyu/yzt/serviceimpl/PsUserService.java
@@ -9,9 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Map;
-
 
 @Service
 @Transactional(rollbackFor=Exception.class)
@@ -25,7 +22,7 @@ public class PsUserService extends ServiceImpl<PsUserMapper,PsUser> implements I
         Integer res;
         QueryWrapper<PsUser> wrapper = new QueryWrapper<>();
         wrapper.eq("openid", user.getOpenid());
-        Integer count = mapper.selectCount(wrapper);
+        Long count = mapper.selectCount(wrapper);
         if (count > 0) {
             res = mapper.updateById(user);
         } else {
